+++ src/questionnaire/web/model/Option.java
@@ -1,11 +1,13 @@
 package questionnaire.web.model;
 
+import java.math.BigDecimal;
+
 public class Option {
 
 	private String optionID;
 	private String questionID;
 	private String caption;
-	private int score;
+	private BigDecimal score;
 	private int sort;
 
 	public String getOptionID() {
@@ -32,11 +34,11 @@ public class Option {
 		this.caption = caption;
 	}
 
-	public int getScore() {
+	public BigDecimal getScore() {
 		return score;
 	}
 
-	public void setScore(int score) {
+	public void setScore(BigDecimal score) {
 		this.score = score;
 	}
